requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
asyncio
aiofiles>=23.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
notebook>=6.5.0 